Restore ES module and add filter to remove method

+++ src/utils/remove.js
@@ -1,8 +1,6 @@
-'use strict';
-
-const fs = require('fs');
-const { join } = require('path');
-const { promisify } = require('util');
+import fs from 'fs';
+import { join } from 'path';
+import { promisify } from 'util';
 
 
 const lstat = promisify(fs.lstat);
@@ -11,7 +9,7 @@ const rmdir = promisify(fs.rmdir);
 const unlink = promisify(fs.unlink);
 
 
-async function _remove(path) {
+async function _remove(path, filter) {
 	try {
 		const stat = await lstat(path);
 
@@ -21,12 +19,14 @@ async function _remove(path) {
 			for (const entity of entities) {
 				const subpath = join(path, entity);
 
-				await _remove(subpath);
+				await _remove(subpath, filter);
 			}
 
-			await rmdir(path);
+			if (!filter || filter(path)) {
+				await rmdir(path);
+			}
 		}
-		else {
+		else if (!filter || filter(path)) {
 			await unlink(path);
 		}
 	}
@@ -44,12 +44,24 @@ async function _remove(path) {
  * Method removes file or folder with it's content.
  * @async @method remove
  * @param {string} path - path to remove
+ * @param {function|regexp} [filter] - remove if function or regexp test returns true
  * @returns {void}
  */
-module.exports.remove = async function (path) {
+export async function remove(path, filter) {
+	let removeFilter = filter;
+
 	if (typeof path !== 'string') {
 		throw new TypeError('"path" is not a string');
 	}
 
-	return await _remove(path);
-};
+	if (filter !== undefined) {
+		if (filter instanceof RegExp) {
+			removeFilter = (src) => filter.test(src);
+		}
+		else if (typeof filter !== 'function') {
+			throw new TypeError('"filter" is not a function nor regexp');
+		}
+	}
+
+	return await _remove(path, removeFilter);
+}
